phieu nhap ne (chua hoan thanh nha may ma)

+++ src/main/java/com/example/demo/model/PhieuNhap.java
@@ -17,6 +17,6 @@ public class PhieuNhap {
 
     private String mancc;
 
-    private Integer tongtien;
+    private Double tongtien;
 
 }
